server/requirements.txt
fastapi==0.115.12
alembic==1.16.1
pydantic_settings==2.9.1
python-jose==3.5.0
passlib==1.7.4
uvicorn==0.34.3
email-validator==2.2.0
asyncpg==0.30.0
